Keep exploration URL handling modular
